Deleted tc-event-mgr-model.c and added files required to generate it #18
